model-server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.4
torch==2.2.0
torchvision==0.17.0
opencv-python==4.9.0.80
python-dotenv==1.0.1 